Added tests
* fixed platform bugs
* fixed v31 expo bugs
