+++ __tests__/AR.mock.js
@@ -1,0 +1,11 @@
+export default {
+  isAvailable() {
+    return false;
+  },
+  removeAllListeners() {},
+  async stopAsync() {},
+  async startAsync() {},
+  getUnavailabilityReason() {
+    return '';
+  },
+};
