+++ __tests__/GLView.mock.js
@@ -1,0 +1,4 @@
+import React from 'react';
+import { View } from 'react-native';
+
+export default View;
