+++ lib/ExpoGraphics.js
@@ -1,5 +1,5 @@
-export { default as View } from './View';
-export { default as GraphicsView } from './View';
+export { default as View } from './GraphicsView';
+export { default as GraphicsView } from './GraphicsView';
 export { default as GLView } from './GLView';
 export { default as ARRunningState } from './ARRunningState';
 export { default as ARCameraState } from './ARCameraState';
