+++ lib/GLView.js
@@ -1,10 +1,10 @@
 // @flow
 import { GLView as EXGLView } from 'expo-gl';
+import invariant from 'invariant';
 import React from 'react';
 import { PixelRatio } from 'react-native';
 
-const invariant = require('fbjs/lib/invariant');
-
+// import EXGLView from '../__tests__/GLView.mock';
 export default class GLView extends React.Component {
   render() {
     const { onContextCreate, ...props } = this.props;
