+++ lib/GraphicsView.ios.js
@@ -1,24 +1,17 @@
 // @flow
 import { AR } from 'expo';
+import invariant from 'invariant';
 import React from 'react';
-import {
-  AppState,
-  findNodeHandle,
-  PixelRatio,
-  Platform,
-  StyleSheet,
-  Text,
-  View,
-} from 'react-native';
+import { AppState, findNodeHandle, PixelRatio, StyleSheet, Text, View } from 'react-native';
 import uuidv4 from 'uuid/v4';
 
-import ARCameraState from '../ARCameraState';
-import ARRunningState from '../ARRunningState';
-import GLView from '../GLView';
+import ARCameraState from './ARCameraState';
+import ARRunningState from './ARRunningState';
+import GLView from './GLView';
 
-const invariant = require('fbjs/lib/invariant');
-
-import type { TrackingConfiguration } from 'expo';
+// import AR from '../__tests__/AR.mock';
+// import type { TrackingConfiguration } from 'expo';
+type TrackingConfiguration = any;
 
 type Layout = {
   x: number,
@@ -42,6 +35,7 @@ type Props = {
   onResize?: (layout: Layout) => void,
   shouldIgnoreSafeGuards?: ?boolean,
 } & React.ElementProps<typeof GLView>;
+
 export default class GraphicsView extends React.Component<Props> {
   nativeRef: ?GLView.NativeView;
   gl: ?any;
@@ -50,7 +44,6 @@ export default class GraphicsView extends React.Component<Props> {
     arRunningProps: {},
     arCameraProps: {},
     isShadowsEnabled: false,
-    arTrackingConfiguration: AR.TrackingConfiguration.World,
     isArCameraStateEnabled: true,
   };
 
@@ -156,16 +149,23 @@ export default class GraphicsView extends React.Component<Props> {
       arTrackingConfiguration,
     } = this.props;
 
-    invariant(onRender, 'ExpoGraphics.View: `onRender` must be defined.');
-    invariant(onContextCreate, 'ExpoGraphics.View: `onContextCreate` must be defined.');
+    invariant(
+      onRender,
+      'expo-graphics: GraphicsView.onContextCreate(): `onRender` must be defined.'
+    );
+    invariant(
+      onContextCreate,
+      'expo-graphics: GraphicsView.onContextCreate(): `onContextCreate` must be defined.'
+    );
 
-    if (Platform.OS === 'ios' && isShadowsEnabled === true) {
+    if (isShadowsEnabled === true) {
       // Bacon: On iOS the Render Buffer must be shimmed to allow for shadows. This will break composers and effects though.
       gl.createRenderbuffer = () => ({});
     }
 
     if (isArEnabled === true) {
-      const trackingConfiguration = arTrackingConfiguration || AR.TrackingConfiguration.World;
+      const { TrackingConfiguration = {} } = AR;
+      const trackingConfiguration = arTrackingConfiguration || TrackingConfiguration.World;
       // Start AR session
       await AR.startAsync(findNodeHandle(this.nativeRef), trackingConfiguration);
     }
