+++ lib/index.js
@@ -1,4 +1,4 @@
 // Makes every `import` variation work
-export * from "./ExpoGraphics";
-import * as ExpoGraphics from "./ExpoGraphics";
+export * from './ExpoGraphics';
+import * as ExpoGraphics from './ExpoGraphics';
 export default ExpoGraphics;
