+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "expo-graphics",
-  "version": "1.0.0-alpha.6",
+  "version": "1.0.1",
   "description": "Tools for building graphic based components in Expo",
   "author": "Evan Bacon <bacon@expo.io> (https://expo.io/)",
   "files": [
@@ -21,11 +21,23 @@
     "opengl"
   ],
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "prettier": "eslint . --fix",
-    "lint:example": "eslint example/",
-    "lint": "eslint lib/",
-    "lint:fix": "eslint lib/ --fix"
+    "jest": "jest",
+    "test-update-snapshot": "jest --updateSnapshot",
+    "lint": "eslint .",
+    "format": "eslint --fix .",
+    "precommit": "lint-staged",
+    "test": "npm run lint && npm run jest"
+  },
+  "lint-staged": {
+    "*.js": [
+      "eslint --fix",
+      "git add"
+    ]
+  },
+  "jest": {
+    "notify": true,
+    "preset": "jest-expo",
+    "testRegex": "/__tests__/[^/]+-test\\.js$"
   },
   "repository": {
     "type": "git",
@@ -34,33 +46,45 @@
   "bugs": {
     "url": "https://github.com/expo/expo-graphics/issues"
   },
-  "peerDependencies": {
-    "react-native": "*",
-    "uuid": "*"
-  },
   "directories": {
     "example": "example",
     "lib": "lib"
   },
+  "peerDependencies": {
+    "react": "*",
+    "react-native": "*"
+  },
   "dependencies": {
-    "expo-gl": "^1.0.1"
+    "expo": "^32.0.0",
+    "expo-gl": "^2.0.0"
   },
   "devDependencies": {
-    "babel-jest": "^23.0.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-expo": "^4.0.0",
-    "babel-preset-react-native": "^4.0.0",
-    "eslint": "^5.1.0",
-    "eslint-config-react-native-prettier": "^1.0.1",
-    "eslint-config-universe": "^1.0.6",
-    "eslint-plugin-jest": "^21.2.0",
-    "flow-bin": "^0.26.0",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.25.0",
+    "babel-eslint": "^7.2.3",
+    "babel-jest": "^22.4.1",
+    "babel-preset-react-native": "^2.1.0",
+    "codecov": "^2.2.0",
+    "conventional-changelog-cli": "^2.0.5",
+    "eslint": "^4.2.0",
+    "eslint-config-prettier": "^2.9.0",
+    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-prettier": "^2.6.0",
+    "eslint-plugin-react": "^7.1.0",
     "husky": "^0.14.3",
-    "jest": "^23.4.1",
-    "jest-expo": "^29.0.0",
-    "prettier": "^1.13.5"
+    "jest": "^22.1.3",
+    "jest-expo": "^25.1.0",
+    "lint-staged": "^4.2.1",
+    "prettier": "^1.12.1",
+    "prettier-eslint": "^8.8.1",
+    "react": "16.2.0",
+    "react-native": "^0.52.0",
+    "react-native-vector-icons": "^4.2.0",
+    "react-test-renderer": "^16.0.0",
+    "release-it": "^7.6.1"
   },
   "eslintConfig": {
-    "extends": "universe/native"
+    "extends": "expo"
   }
 }
